requirements.txt
Django==1.11.12
docker-py==1.10.6       # OSI Approved  Apache Software License
kubernetes=6.0.0
pyyaml
uuid
